requirements.txt
torch>=2.4.0
torchaudio
torchvision
datasets>=3.0.0
streamlit>=1.38.0
wandb>=0.19.0
tqdm
numpy
matplotlib
seaborn
pyyaml